Added markdown formatting + additional info
I added markdown formatting to the text here to make it more easily readable.
I also added a section at the bottom, which was not included as part of my original assignment submission. This "file info" section details the purpose of each of the source files, as their names are not standardized nor obvious to understand.